django-base/requirements.txt
Django==3.0.6
django-import-export
